rag-core/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-huggingface
beautifulsoup4
playwright
pydantic
requests
httpx
sentence-transformers
pypdf
python-docx
python-multipart
pillow
aiofiles
upstash-vector
